requirements/base.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
PyMuPDF==1.23.8
Pillow==10.2.0
numpy==1.26.3
sentence-transformers==2.3.1
FlagEmbedding==1.2.5
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
faiss-cpu==1.7.4
rapidocr-onnxruntime==1.3.24
